+++ ffn_dl/__init__.py
@@ -2,7 +2,7 @@
 
 import logging as _logging
 
-__version__ = "1.20.6"
+__version__ = "1.20.7"
 
 _logger = _logging.getLogger(__name__)